backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
pytesseract==0.3.10
opencv-python-headless==4.9.0.80
Pillow==10.2.0
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
